Revised for Creators Update Added Visual Studio 2017 solution and project files Added support for Windows SDK version 15063 Added support for HDR displays (requires 15063) Replaced tone mapping operators with the newly-devised "Stanard" tone curve Newly revised temporal antialiasing implementation compares depths to detect disocclusion Double-buffered linear depth for temporal algorithms TemporalAA now finishes with a sharpening pass Bloom resolution increased for large resolutions (> 2560x1440) Added ability to readback textures into main memory and write to disk (for screenshots, etc.) Added ability to load texture images from disk into memory Relocated SetStablePowerState() to device creation Detect Developer Mode by reading Windows registry before calling SetStablePowerState() Added more magnification levels to pixel zoom feature Various code improvements and minor bug fixes

+++ MiniEngine/Core/Shaders/AdaptExposureCS.hlsl
@@ -27,7 +27,7 @@ cbuffer cb0 : register(b1)
 	float AdaptationRate;
 	float MinExposure;
 	float MaxExposure;
-	float PixelCount; 
+	uint PixelCount; 
 }
 
 groupshared float gs_Accum[256];
